Small Sound Implementation
Now with 100% more REWARDS FOR PLAYING with the Ball! Sound plays on impact.
AI GUI added to the screen for easy debug + a quick view into the mind of the creature.
Changed "Stomach" GUI element to "Stuffed" as it sounds better!
Changed Ball texture to something more "interesting" LOL
Changed Background sound of the game from SNES Secret of Mana Music to OCRemix Videogame Remixes File